+++ src/utils.rs
@@ -3,7 +3,8 @@ use std::io::{Error, Write};
 use flexi_logger::{AdaptiveFormat, DeferredNow, style};
 use log::Record;
 
-pub const ADAPTIVE_LOG_FORMAT: AdaptiveFormat = AdaptiveFormat::Custom(default_format, colored_default_format);
+pub const ADAPTIVE_LOG_FORMAT: AdaptiveFormat =
+    AdaptiveFormat::Custom(default_format, colored_default_format);
 
 pub fn syslog_prefixed_format(
     w: &mut dyn Write,
@@ -18,10 +19,10 @@ pub fn syslog_prefixed_format(
         log::Level::Debug | log::Level::Trace => "<7>",
     };
 
-    if let Some(m) = record.module_path() {
-        if m.starts_with("thumbs") {
-            return write!(w, "{prefix}{}", record.args());
-        }
+    if let Some(m) = record.module_path()
+        && m.starts_with("thumbs")
+    {
+        return write!(w, "{prefix}{}", record.args());
     }
 
     write!(
@@ -39,10 +40,10 @@ pub fn default_format(
     record: &Record,
 ) -> Result<(), Error> {
     let level = record.level();
-    if let Some(m) = record.module_path() {
-        if m.starts_with("thumbs") {
-            return write!(w, "{} {}", level, record.args());
-        }
+    if let Some(m) = record.module_path()
+        && m.starts_with("thumbs")
+    {
+        return write!(w, "{} {}", level, record.args());
     }
 
     write!(
@@ -62,15 +63,15 @@ pub fn colored_default_format(
 ) -> Result<(), Error> {
     let level = record.level();
 
-    if let Some(m) = record.module_path() {
-        if m.starts_with("thumbs") {
-            return write!(
-                w,
-                "{} {}",
-                style(level).paint(level.to_string()),
-                style(level).paint(record.args().to_string())
-            );
-        }
+    if let Some(m) = record.module_path()
+        && m.starts_with("thumbs")
+    {
+        return write!(
+            w,
+            "{} {}",
+            style(level).paint(level.to_string()),
+            style(level).paint(record.args().to_string())
+        );
     }
 
     write!(